Manage prompts added to non-managed groups
If an integer ID of a group is used in Uiprompt:new, it will be assumed
the group is not managed and the prompt will instead of managed as if it
were a standalone prompt.

This also makes the old UipromptGroup:addPrompt obsolete and replaces it
with a new Uiprompt:addPrompt method.